removed website
due to a sudden political change in our school, another website made another class was banned due to "It's inappropriate" (Teacher). And the discovery of this website is only a matter of time, so I will delete all files from this repository and keep only a zip file. For my own good and for our class.